requirements.txt
easyocr
opencv-python
numpy
Pillow
pdf2image
python-docx